refactor, fix: change way of providing configuration, add utilities, fix providing role claims

+++ Invoicing/Invoicing.Identity.API/Seeders/DbSeeder.cs
@@ -3,18 +3,28 @@ using Duende.IdentityServer.EntityFramework.DbContexts;
 using Duende.IdentityServer.EntityFramework.Mappers;
 using IdentityModel;
 using Invoicing.Identity.API.Configuration;
+using Invoicing.Identity.API.Configuration.Interfaces.Interfaces;
+using Invoicing.Identity.API.Extensions;
 using Invoicing.Identity.Domain.Entities;
 using Invoicing.Identity.Domain.Enums;
 using Invoicing.Identity.Infrastructure.Data;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Options;
 using Serilog;
 
 namespace Invoicing.Identity.API.Seeders;
 
-public class Seeder
+public class DbSeeder : IDbSeeder
 {
-    public static async Task EnsureSeedData(WebApplication app)
+    private readonly IOptions<AppSettings> _appSettings;
+
+    public DbSeeder(IOptions<AppSettings> appSettings)
+    {
+        _appSettings = appSettings ?? throw new ArgumentException(nameof(appSettings));
+    }
+
+    public async Task EnsureSeedData(WebApplication app)
     {
         using (var scope = app.Services.GetRequiredService<IServiceScopeFactory>().CreateScope())
         {
@@ -25,53 +35,65 @@ public class Seeder
             var applicationDbContext = scope.ServiceProvider.GetService<ApplicationDbContext>() ??
                                        throw new ArgumentNullException(nameof(ApplicationDbContext));
 
-            await persistedGrantDbContext.Database.EnsureDeletedAsync();
-            await configurationDbContext.Database.EnsureDeletedAsync();
-            await applicationDbContext.Database.EnsureDeletedAsync();
+            if (app.Environment.IsSystemTests())
+            {
+                await persistedGrantDbContext.Database.EnsureDeletedAsync();
+                await configurationDbContext.Database.EnsureDeletedAsync();
+                await applicationDbContext.Database.EnsureDeletedAsync();
+            }
+
             await persistedGrantDbContext.Database.MigrateAsync();
             await configurationDbContext.Database.MigrateAsync();
             await applicationDbContext.Database.MigrateAsync();
 
             var roleManager = scope.ServiceProvider.GetRequiredService<RoleManager<IdentityRole>>();
             var userManager = scope.ServiceProvider.GetRequiredService<UserManager<ApplicationUser>>();
+            var defaultEntitiesProvider = scope.ServiceProvider.GetRequiredService<IDefaultEntitiesProvider>();
+            var defaultIdentityConfigProvider = scope.ServiceProvider.GetRequiredService<IDefaultIdentityConfigurationProvider>();
 
             await SeedRoles(roleManager);
-            await SeedUsers(userManager);
-            await SeedConfig(configurationDbContext);
+            await SeedUsers(userManager, defaultEntitiesProvider);
+            await SeedConfig(configurationDbContext, defaultIdentityConfigProvider);
+
+            await persistedGrantDbContext.SaveChangesAsync();
+            await configurationDbContext.SaveChangesAsync();
+            await applicationDbContext.SaveChangesAsync();
         }
     }
 
-    private static async Task SeedConfig(ConfigurationDbContext configurationDbContext)
+    private async Task SeedConfig(ConfigurationDbContext configurationDbContext,
+        IDefaultIdentityConfigurationProvider defaultIdentityConfigurationProvider)
     {
-        await AddIdentityClients(configurationDbContext);
-        await AddIdentityApiScopes(configurationDbContext);
-        await AddIdentityResources(configurationDbContext);
-
-        await configurationDbContext.SaveChangesAsync();
+        await AddIdentityClients(configurationDbContext, defaultIdentityConfigurationProvider);
+        await AddIdentityApiScopes(configurationDbContext, defaultIdentityConfigurationProvider);
+        await AddIdentityResources(configurationDbContext, defaultIdentityConfigurationProvider);
     }
 
-    private static async Task AddIdentityClients(ConfigurationDbContext configurationDbContext)
+    private async Task AddIdentityClients(ConfigurationDbContext configurationDbContext,
+        IDefaultIdentityConfigurationProvider identityConfigurationProvider)
     {
         if (!configurationDbContext.Clients.Any())
             await configurationDbContext.Clients.AddRangeAsync(
-                IdentityConfig.Clients.Select(client => client.ToEntity()));
+                identityConfigurationProvider.GetClients.Select(client => client.ToEntity()));
     }
 
-    private static async Task AddIdentityResources(ConfigurationDbContext configurationDbContext)
+    private async Task AddIdentityResources(ConfigurationDbContext configurationDbContext,
+        IDefaultIdentityConfigurationProvider identityConfigurationProvider)
     {
         if (!configurationDbContext.IdentityResources.Any())
             await configurationDbContext.IdentityResources.AddRangeAsync(
-                IdentityConfig.IdentityResources.Select(res => res.ToEntity()));
+                identityConfigurationProvider.GetIdentityResources.Select(res => res.ToEntity()));
     }
 
-    private static async Task AddIdentityApiScopes(ConfigurationDbContext configurationDbContext)
+    private async Task AddIdentityApiScopes(ConfigurationDbContext configurationDbContext,
+        IDefaultIdentityConfigurationProvider identityConfigurationProvider)
     {
         if (!configurationDbContext.ApiScopes.Any())
             await configurationDbContext.ApiScopes.AddRangeAsync(
-                IdentityConfig.ApiScopes.Select(apiScope => apiScope.ToEntity()));
+                identityConfigurationProvider.GetApiScopes.Select(apiScope => apiScope.ToEntity()));
     }
 
-    private static async Task SeedRoles(RoleManager<IdentityRole> roleManager)
+    private async Task SeedRoles(RoleManager<IdentityRole> roleManager)
     {
         IEnumerable<string> defaultRoles = Enum.GetNames(typeof(Roles));
         foreach (var role in defaultRoles)
@@ -84,9 +106,9 @@ public class Seeder
         }
     }
 
-    private static async Task SeedUsers(UserManager<ApplicationUser> userManager)
+    private async Task SeedUsers(UserManager<ApplicationUser> userManager, IDefaultEntitiesProvider defaultEntitiesProvider)
     {
-        foreach ((var role, var users) in DefaultEntities.DefaultUsers)
+        foreach (var (role, users) in defaultEntitiesProvider.DefaultUsers)
         foreach (var applicationUser in users)
         {
             if (!await DoesUserExistAsync(userManager, applicationUser)) continue;
@@ -99,7 +121,7 @@ public class Seeder
         }
     }
 
-    private static async Task<bool> DoesUserExistAsync(UserManager<ApplicationUser> userManager, ApplicationUser applicationUser)
+    private async Task<bool> DoesUserExistAsync(UserManager<ApplicationUser> userManager, ApplicationUser applicationUser)
     {
         var userInDatabase = await userManager.FindByNameAsync(applicationUser.UserName);
 
@@ -112,21 +134,22 @@ public class Seeder
         return true;
     }
 
-    private static async Task CreateUserAsync(UserManager<ApplicationUser> userManager, ApplicationUser applicationUser)
+    private async Task CreateUserAsync(UserManager<ApplicationUser> userManager, ApplicationUser applicationUser)
     {
-        var createUserResult = await userManager.CreateAsync(applicationUser, "defaultPassword123!");
+        var createUserResult = await userManager.CreateAsync(applicationUser,
+            _appSettings.Value.DefaultUserPassword ?? throw new ArgumentNullException(_appSettings.Value.DefaultUserPassword));
 
         if (!createUserResult.Succeeded) RaiseIdentityError(createUserResult);
     }
 
-    private static async Task AssignToRoleAsync(UserManager<ApplicationUser> userManager, ApplicationUser applicationUser, Roles role)
+    private async Task AssignToRoleAsync(UserManager<ApplicationUser> userManager, ApplicationUser applicationUser, Roles role)
     {
         var addRolesResult = await userManager.AddToRoleAsync(applicationUser, role.ToString());
 
         if (!addRolesResult.Succeeded) RaiseIdentityError(addRolesResult);
     }
 
-    private static async Task AddClaimsAsync(UserManager<ApplicationUser> userManager, ApplicationUser applicationUser, params Claim[] claims)
+    private async Task AddClaimsAsync(UserManager<ApplicationUser> userManager, ApplicationUser applicationUser, params Claim[] claims)
     {
         var defaultClaims = new List<Claim>
         {
@@ -138,14 +161,14 @@ public class Seeder
         if (!addClaimsResult.Succeeded) RaiseIdentityError(addClaimsResult);
     }
 
-    private static void RaiseIdentityError(IdentityResult failedResult)
+    private void RaiseIdentityError(IdentityResult failedResult)
     {
         var message = $"Identity operation failed: {GetErrorsAsString(failedResult)}";
         Log.Error(message);
         throw new Exception(message);
     }
 
-    private static string GetErrorsAsString(IdentityResult result)
+    private string GetErrorsAsString(IdentityResult result)
     {
         return string.Join(Environment.NewLine, result.Errors.Select(x => x.Description));
     }
